Correct fishing values
It came to our attention that some of the  calculator values may not be correct. So I looked at each value for tourism, fishing and CV in this matrix, and compared with the colors on the game board. All of the values were correct, except one consistent problem with fishing - all of the 1 values should be 2, and all of the 2 values should be 1. So this has been corrected.